requirements.txt
absl_py==2.2.2
diff_gaussian_rasterization==0.0.0
fused_ssim==0.0.0
fused_ssim.egg==info
joblib==1.4.2
matplotlib==3.10.3
natsort==8.4.0
numpy==2.2.6
opencv_python==4.10.0.84
Pillow==11.2.1
plyfile==1.1.2
pytorch_msssim==1.0.0
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.3
setuptools==75.1.0
simple_knn==0.0.0
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.0
